requirements.txt
Flask==3.1.2
Werkzeug==3.1.3
python-dotenv==1.2.1
PyYAML==6.0.3
google-generativeai==0.7.0
markdown2==2.5.4
reportlab==4.0.9
